// workspace/src/lib/data/types.ts
export type JenisVaksin =
  | 'BCG'
  | 'Hepatitis B 0'
  | 'Polio 1'
  | 'Polio 2'
  | 'Polio 3'
  | 'Polio 4'
  | 'DPT-HB-Hib 1'
  | 'DPT-HB-Hib 2'
  | 'DPT-HB-Hib 3'
  | 'Campak'
  | 'MR'
  | 'PCV'
  | 'Rotavirus';

export type StatusImunisasi = 'Lengkap' | 'Belum Lengkap' | 'Terlambat' | 'Belum Dimulai';